Add missing docs and remove a field
Also convert the Layout field in the card face from a string to a Layout
type.

The flavor_name field doesn't seem to be documented on the card face,
only the card object. If we find a card that has this set on the card
face then we can add it back.